post-delete

+++ modules/UshahidiUI/media/js/app/views/PostItemView.js
@@ -1,5 +1,5 @@
-define(['App', 'marionette', 'underscore', 'handlebars', 'text!templates/PostListItem.html'],
-	function(App, Marionette, _, Handlebars, template)
+define(['App', 'marionette', 'underscore', 'handlebars', 'alertify', 'text!templates/PostListItem.html'],
+	function(App, Marionette, _, Handlebars, alertify, template)
 	{
 		//ItemView provides some default rendering logic
 		return Marionette.ItemView.extend(
@@ -18,10 +18,20 @@ define(['App', 'marionette', 'underscore', 'handlebars', 'text!templates/PostLis
 			// @todo add confirmation dialog
 			deletepost: function(e)
 			{
-				e.preventDefault();
-				this.model.destroy({
-					// Wait till server responds before destroying model
-					wait: true
+				alertify.confirm("Are you sure you want to delete", function(e)
+				{
+					e.preventDefault();
+					this.model.destroy({
+						// Wait till server responds before destroying model
+						wait: true
+					});
+
+					if(true){
+						alertify.success("Post has been deleted.");
+
+					} else {
+						alertify.error("Please try again");
+					}
 				});
 			},
